app/core/scheduler.py
import asyncio
from typing import Dict, Any, List, Callable
from datetime import datetime, timedelta
import schedule
import time
import threading
from concurrent.futures import ThreadPoolExecutor

from app.utils.logger import get_logger
from app.core.config import settings
from app.data_collectors import (
    CoinGeckoCollector,
    YahooFinanceCollector,
    FredCollector,
    SentimentCollector
)


class DataCollectionScheduler:
    """Scheduler for automated data collection tasks."""

    def __init__(self):
        self.logger = get_logger("scheduler")
        self.is_running = False
        self.executor = ThreadPoolExecutor(max_workers=4)

        # Initialize collectors
        self.collectors = {
            "coingecko": CoinGeckoCollector(),
            "yahoo_finance": YahooFinanceCollector(),
            "fred": FredCollector(),
            "sentiment": SentimentCollector()
        }

        self.setup_schedules()

    def setup_schedules(self) -> None:
        """Set up collection schedules for different data sources."""
        try:
            # Cryptocurrency data - every hour
            schedule.every().hour.do(
                lambda: self.executor.submit(self.collect_crypto_data)
            )

            # Traditional market data - every 30 minutes during market hours
            schedule.every(30).minutes.do(
                lambda: self.executor.submit(self.collect_market_data)
            )

            # FRED economic data - daily at 9 AM
            schedule.every().day.at("09:00").do(
                lambda: self.executor.submit(self.collect_economic_data)
            )

            # Sentiment data - every 4 hours
            schedule.every(4).hours.do(
                lambda: self.executor.submit(self.collect_sentiment_data)
            )

            # Correlation calculations - daily at 10 PM
            schedule.every().day.at("22:00").do(
                lambda: self.executor.submit(self.calculate_correlations)
            )

            # Risk metrics - daily at 11 PM
            schedule.every().day.at("23:00").do(
                lambda: self.executor.submit(self.calculate_risk_metrics)
            )

            # Weekly report generation - Sundays at 8 AM
            schedule.every().sunday.at("08:00").do(
                lambda: self.executor.submit(self.generate_weekly_report)
            )

            self.logger.info("Data collection schedules configured successfully")

        except Exception as e:
            self.logger.error("Failed to setup schedules", error=str(e))

    def collect_crypto_data(self) -> Dict[str, Any]:
        """Collect cryptocurrency data."""
        try:
            self.logger.info("Starting crypto data collection")

            collector = self.collectors["coingecko"]
            data = collector.collect_data()

            if data:
                # Store to database
                # TODO: Implement database storage
                self.logger.info("Crypto data collected successfully", records=len(data.get("market_data", [])))
                return {"status": "success", "data": data}
            else:
                self.logger.warning("No crypto data collected")
                return {"status": "failed", "error": "No data received"}

        except Exception as e:
            self.logger.error("Failed to collect crypto data", error=str(e))
            return {"status": "error", "error": str(e)}

    def collect_market_data(self) -> Dict[str, Any]:
        """Collect traditional market data."""
        try:
            self.logger.info("Starting market data collection")

            collector = self.collectors["yahoo_finance"]
            data = collector.collect_data()

            if data:
                # Store to database
                # TODO: Implement database storage
                self.logger.info("Market data collected successfully", records=len(data.get("market_data", [])))
                return {"status": "success", "data": data}
            else:
                self.logger.warning("No market data collected")
                return {"status": "failed", "error": "No data received"}

        except Exception as e:
            self.logger.error("Failed to collect market data", error=str(e))
            return {"status": "error", "error": str(e)}

    def collect_economic_data(self) -> Dict[str, Any]:
        """Collect economic indicators from FRED."""
        try:
            self.logger.info("Starting economic data collection")

            collector = self.collectors["fred"]
            data = collector.collect_data()

            if data:
                # Store to database
                # TODO: Implement database storage
                self.logger.info("Economic data collected successfully", records=len(data.get("fred_data", [])))
                return {"status": "success", "data": data}
            else:
                self.logger.warning("No economic data collected")
                return {"status": "failed", "error": "No data received"}

        except Exception as e:
            self.logger.error("Failed to collect economic data", error=str(e))
            return {"status": "error", "error": str(e)}

    def collect_sentiment_data(self) -> Dict[str, Any]:
        """Collect sentiment indicators."""
        try:
            self.logger.info("Starting sentiment data collection")

            collector = self.collectors["sentiment"]
            data = collector.collect_data()

            if data:
                # Store to database
                # TODO: Implement database storage
                self.logger.info("Sentiment data collected successfully")
                return {"status": "success", "data": data}
            else:
                self.logger.warning("No sentiment data collected")
                return {"status": "failed", "error": "No data received"}

        except Exception as e:
            self.logger.error("Failed to collect sentiment data", error=str(e))
            return {"status": "error", "error": str(e)}

    def calculate_correlations(self) -> Dict[str, Any]:
        """Calculate asset correlations."""
        try:
            self.logger.info("Starting correlation calculations")

            # TODO: Implement correlation calculations using stored data
            # This would fetch price data from database and calculate correlations

            self.logger.info("Correlation calculations completed")
            return {"status": "success", "message": "Correlations calculated"}

        except Exception as e:
            self.logger.error("Failed to calculate correlations", error=str(e))
            return {"status": "error", "error": str(e)}

    def calculate_risk_metrics(self) -> Dict[str, Any]:
        """Calculate risk metrics for all assets."""
        try:
            self.logger.info("Starting risk metrics calculations")

            # TODO: Implement risk metrics calculations using stored data
            # This would fetch price data and calculate VaR, volatility, etc.

            self.logger.info("Risk metrics calculations completed")
            return {"status": "success", "message": "Risk metrics calculated"}

        except Exception as e:
            self.logger.error("Failed to calculate risk metrics", error=str(e))
            return {"status": "error", "error": str(e)}

    def generate_weekly_report(self) -> Dict[str, Any]:
        """Generate weekly analysis report."""
        try:
            self.logger.info("Starting weekly report generation")

            # TODO: Implement weekly report generation
            # This would analyze the week's data and generate insights

            self.logger.info("Weekly report generated successfully")
            return {"status": "success", "message": "Weekly report generated"}

        except Exception as e:
            self.logger.error("Failed to generate weekly report", error=str(e))
            return {"status": "error", "error": str(e)}

    def check_and_send_alerts(self) -> Dict[str, Any]:
        """Check for alert conditions and send notifications."""
        try:
            self.logger.info("Checking for alert conditions")

            alerts_sent = []

            # TODO: Implement alert checking logic
            # - Check correlation breakdowns
            # - Check volatility spikes
            # - Check liquidity changes

            if alerts_sent:
                self.logger.info(f"Sent {len(alerts_sent)} alerts")
            else:
                self.logger.info("No alerts triggered")

            return {"status": "success", "alerts_sent": len(alerts_sent)}

        except Exception as e:
            self.logger.error("Failed to check alerts", error=str(e))
            return {"status": "error", "error": str(e)}

    def run_scheduler(self) -> None:
        """Run the scheduler in a loop."""
        self.is_running = True
        self.logger.info("Data collection scheduler started")

        try:
            while self.is_running:
                schedule.run_pending()
                time.sleep(60)  # Check every minute

        except KeyboardInterrupt:
            self.logger.info("Scheduler stopped by user")
        except Exception as e:
            self.logger.error("Scheduler error", error=str(e))
        finally:
            self.stop_scheduler()

    def stop_scheduler(self) -> None:
        """Stop the scheduler."""
        self.is_running = False
        self.executor.shutdown(wait=True)
        self.logger.info("Data collection scheduler stopped")

    def start_background_scheduler(self) -> threading.Thread:
        """Start scheduler in background thread."""
        scheduler_thread = threading.Thread(target=self.run_scheduler, daemon=True)
        scheduler_thread.start()
        return scheduler_thread

    def get_schedule_status(self) -> Dict[str, Any]:
        """Get current schedule status."""
        return {
            "is_running": self.is_running,
            "scheduled_jobs": len(schedule.jobs),
            "next_runs": [
                {
                    "job": str(job.job_func),
                    "next_run": job.next_run.isoformat() if job.next_run else None
                }
                for job in schedule.jobs
            ]
        }

    def run_manual_collection(self, collector_name: str) -> Dict[str, Any]:
        """Run a specific collector manually."""
        try:
            if collector_name == "crypto":
                return self.collect_crypto_data()
            elif collector_name == "market":
                return self.collect_market_data()
            elif collector_name == "economic":
                return self.collect_economic_data()
            elif collector_name == "sentiment":
                return self.collect_sentiment_data()
            else:
                return {"status": "error", "error": f"Unknown collector: {collector_name}"}

        except Exception as e:
            self.logger.error(f"Failed to run manual collection for {collector_name}", error=str(e))
            return {"status": "error", "error": str(e)}


# Global scheduler instance
data_scheduler = DataCollectionScheduler()


def start_scheduler() -> None:
    """Start the global data scheduler."""
    data_scheduler.start_background_scheduler()


def stop_scheduler() -> None:
    """Stop the global data scheduler."""
    data_scheduler.stop_scheduler()


def get_scheduler_status() -> Dict[str, Any]:
    """Get scheduler status."""
    return data_scheduler.get_schedule_status()