requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data Collection & APIs
requests==2.31.0
httpx==0.25.2
yfinance==0.2.26
pandas-datareader==0.10.0
fredapi==0.5.1

# Data Processing & Analytics
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0
TA-Lib==0.4.28

# Machine Learning
tensorflow==2.15.0
keras==2.15.0
torch==2.1.2
xgboost==2.0.3

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.17.0
altair==5.2.0
streamlit-aggrid==0.3.4
streamlit-option-menu==0.3.6

# Task Scheduling
celery==5.3.4
python-crontab==3.0.0
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0